refactor(deviceauth): split 861-line sqlite store by entity (core→115)
The SQLite store packed five entities and shared codec into one file.
Split by entity within package store — pure relocation, no call-site or
behavior changes. Each entity file owns its public methods plus its
clone/put/get-tx/scan helpers:

- sqlite.go (115): store struct, ctor, Close, schema migrations
- sqlite_pairing.go / sqlite_device.go / sqlite_grant.go /
  sqlite_session.go / sqlite_bootstrap.go: per-entity CRUD
- sqlite_codec.go: exec/scanner interfaces, capability (un)marshal,
  millis conversion, SQL error normalization

53 funcs in, 53 funcs out (lossless). Verified by go build, go vet,
the deviceauth store tests, scoped golangci-lint, goimports.
